style: 모든 리뷰 페이지 관련 css 수정
